Update Controller.java
Added controls for player1 and player2 paddles as well as initialized p1score and p2score int values for future use. Controls were also with exact bounds that prevent x/y movement in their own respective directions. Due to exact measurements of both the gameArena window and the hockey table, I used both pen and paper as well as Desmos to aide calculations and to guarantee ease of future implementation and symmetry.