Update Controller.java by adding game loop and goal collision
Added goal collisions along with game loop that also works as a robust way of resetting the game by having a boolean value superseding all game logic. However, it is still under the while(true) loop so that the gameArena object window can record the spacePressed() to reset the game. I plan on implementing sound button a similar way.