Added sound class
Added a sound class that uses 'Runnable' to play sounds on a separate thread to prevent the game logic from freezing or slowing down. I have also added a buffer to cut up the input/outputs to smoothen the sound albeit with slight latency. The sound is called by playSound() and will be controlled by a gameArena.enterPressed() and gameArena.shiftPressed() as a boolean for the button. This sound player class only works on the precedent that the sound files are located within the same folder as the Controller.java (which they are).