Update Controller.java
Implemented cheat codes and a basic sound button that sets text at the top to either red or green based on if sound is true or false. All instances of sound are inside if statements based on whether boolean 'sound' is true. The implementation of the cheat codes are done in a similar fashion but without the added boolean variable. Once either 'o' or 'p' is pressed, the next goal to each respective key-player will be the winning goal. Both of these buttons, of course, are within the main game loop.